三つ目です。最後も近日アップします。
r34 == r17_beta16(2010-12-05) 231_r34_kene.zip (kene素材ロダ)

昨日鼻だったので嫌な予感はしてたのですが、本格的に風邪みたいでのどが痛い。

† 幽々子性能変更。
† 1～3面はとりあえず簡単に。
† BAD END追加。easyは5面までになりました。
† ポーズ画面がぼやぼやになった。
† スクショ作り直した。
† 処理落ちとか改善中。(ボム撃つとダメダメですが)
† ライブラリ改善中。
† なんか色々おかしいけど、直す時間なかった。
† その他。

//---------------------------------------------
■７．ＦＡＱ
    Ｑ．次の更新日はいつですか……
    Ａ．更新日について
      更新は一年に一回ぐらいで十分な気もしますが、
      基本的に年４回。つまり春夏秋冬です。季節感を大事にするのですよ(?)。
       1  2  3春
       4  5  6夏
       7  8  9秋
      10 11 12冬
      そんで一週間ぐらい過ぎた始めの日曜日です。(1日が日曜の場合は8日にする)
    2011年は
      夏2011-06-05(日曜日)
      秋2011-09-04(日曜日)
      冬2011-12-04(日曜日)
    になります。(ただし予定は未定です)

    Ｑ．前の奴の方が面白かったんだけど……
    Ａ．そういう場合は、前の奴で遊んで下さい。
      ms0:/PSP/GAME/kene_r34/
      ms0:/PSP/GAME/kene_r33/
      ms0:/PSP/GAME/kene_r32/
        ...
      みたいになってても全然問題ありません。
//---------------------------------------------

次回更新は夏(2011-06-05)頃を予定してます。

+++ trunk/src/core/draw/draw_kanji.c
@@ -418,7 +418,7 @@ global /*static*/ void script_message_window_clear(void)
 
 //	inits				= 1;/*???*/
 //	is_bg				= 0;/*???*/
-//	draw_script_screen	= 0;/*???*/
+//	cg.draw_flag_script_screen	= 0;/*???*/
 
 	home_cursor();		/* カーソルをホームポジションへ移動 */
 //	cursor_x_chached	= 0;/*???*/ /* カーソル初期化 */
@@ -440,7 +440,7 @@ global /*static*/ void script_message_window_clear(void)
 	 (wait指定で、frame毎に一文字描画。 wait0で、続きを全描画)
 ---------------------------------------------------------*/
 
-global /*static*/ int print_kanji000(const char *str, int color_type, int wait)
+global /*static*/ int print_kanji000(const char *str, int color_type, /*(要確認)*/int wait)
 {
 	font_color_number = (color_type & 0x0f);
 //
@@ -449,7 +449,7 @@ global /*static*/ int print_kanji000(const char *str, int color_type, int wait)
 	{	/* wait が 0 の場合は一回で全文字描画する */
 		static unsigned int count_wait_time;
 		count_wait_time++;
-		if (count_wait_time > wait)
+		if (count_wait_time > (unsigned/*(要確認)*/)wait)
 		{
 			count_wait_time = 0;
 			need_draw_this_flame = 1;
